split into two scripts
- original script now handles making queries and formatting reponses
- new script handles CLI args, setup, and outputs
